requirements.txt
langchain==1.0.5
langchain_core==1.0.4
langchain_community>=0.1.0
langchain_text_splitters>=0.1.0
langchain_huggingface>=0.1.0
langchain_ollama>=0.1.0
langchain_openai>=0.1.0
llama_index>=0.5.0
llama-index-core>=0.5.0
llama-index-embeddings-openai
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
langgraph-sdk>=0.2.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
finviz==1.4.6
yfinance>=0.2.31
pydantic>=2.0.0
python-dotenv>=1.0.0
langsmith>=0.0.70
plotly>=5.15.0
seaborn>=0.12.0
ollama>=0.1.0
vllm>=0.3.3
requests
ta
nest_asyncio
dashscope
akshare